Add space between helpers
Extracted from #37

Also reorganized the theme library into more distinct modules
