+++ libs/theme/src/index.ts
@@ -1,3 +1,4 @@
-export * from './lib/theme'
 export * from './lib/colors'
 export * from './lib/global'
+export * from './lib/theme'
+export * from './lib/types'
