+++ libs/theme/src/lib/colors.ts
@@ -1,8 +1,4 @@
-type ColorNames = 'gray' | 'red' | 'yellow' | 'blue' | 'green'
-type ColorValues = 50 | 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900
-
-export type Color = `${ColorNames}${ColorValues}` | 'white' | 'black'
-export type ColorPalette = Record<Color, string>
+import { ColorPalette } from './types'
 
 export const colorPalette: ColorPalette = {
   white: 'hsl(0, 0%, 100%)',
