+++ libs/theme/src/lib/theme/dark.ts
@@ -1,20 +1,4 @@
-import { DefaultTheme as Theme } from 'styled-components'
-import { colorPalette } from '../colors'
-
-const baseTheme: Theme = {
-  // TODO: Host these font files and use @font-face declarations in a global stylesheet
-  fonts: {
-    sans: `'Inter', sans-serif`,
-    mono: `'GT America Mono', monospace`,
-  },
-  themeColors: colorPalette,
-
-  // Spacing is based on Tailwind's default spacing scale. See: https://tailwindcss.com/docs/customizing-spacing#default-spacing-scale
-  // Usage: `theme.spacing(4)` => '1rem' (16px)
-  //
-  // Dev Note: Eventually most commonly used numbers will map to strings e.g. `theme.spacing('small')`
-  spacing: (size) => `${size * 0.25}rem`,
-}
+import { baseTheme } from './base'
 
 export const darkTheme = {
   ...baseTheme,
