+++ libs/theme/src/lib/theme/index.ts
@@ -1,41 +1,4 @@
-import 'styled-components'
-import { ColorPalette } from '../colors'
 import { darkTheme } from './dark'
 
 export * from './dark'
 export const defaultTheme = darkTheme
-
-type SizingMultiplier =
-  | 0.5
-  | 1
-  | 2
-  | 3
-  | 4
-  | 5
-  | 6
-  | 7
-  | 8
-  | 9
-  | 10
-  | 11
-  | 12
-  | 16
-  | 20
-  | 24
-  | 32
-  | 40
-  | 48
-  | 56
-  | 64
-
-// Extend the styled-components
-declare module 'styled-components' {
-  export interface DefaultTheme {
-    themeColors: ColorPalette
-    fonts: {
-      sans: string
-      mono: string
-    }
-    spacing: (size: SizingMultiplier) => string
-  }
-}
